chore: implemented changes in the codebase

+++ nft_staking_q3/programs/nft_staking_q3/src/state/stake_config.rs
@@ -3,7 +3,7 @@ use anchor_lang::prelude::*;
 #[account]
 #[derive(InitSpace)]
 pub struct StakeConfig {
-    pub points_per_stake: u8,
+    pub points_per_stake: i64,
     pub max_staked: u8,
     pub freeze_period: i64,
     pub reward_bump: u8,
